fix(build): resolve SSR issues with PDF viewer component
Fixes "DOMMatrix is not defined" error during Next.js build process
by extracting PDF.js functionality into a client-only component.

Changes:
- Create separate PDFViewer component with all PDF.js dependencies
- Implement dynamic import with ssr: false to prevent server execution
- Add loading state during component hydration
- Maintain all existing functionality (zoom, pagination, fullscreen, download, tooltips)

This resolves production build failures while preserving user experience.
